+++ test/strongbox/github_api_test.clj
@@ -589,7 +589,7 @@
                    :source "github",
                    :source-id "0xbs/premade-applicants-filter",
                    :tag-list [],
-                   :updated-date "2021-12-26T09:40:18+00:00"
+                   :updated-date "2021-12-26T09:40:18Z"
                    :url "https://github.com/0xbs/premade-applicants-filter"}
                   {:download-count 0,
                    :game-track-list [:classic-tbc :classic :retail],
@@ -598,7 +598,7 @@
                    :source "github",
                    :source-id "AlexFolland/ArenaLeaveConfirmer",
                    :tag-list [],
-                   :updated-date "2021-07-04T22:12:06+00:00",
+                   :updated-date "2021-07-04T22:12:06Z",
                    :url "https://github.com/AlexFolland/ArenaLeaveConfirmer"}
                   {:download-count 0,
                    :game-track-list [:classic-tbc :classic :retail],
@@ -607,7 +607,7 @@
                    :source "github",
                    :source-id "AlexFolland/BattlegroundSpiritReleaser",
                    :tag-list [],
-                   :updated-date "2021-07-04T21:55:31+00:00",
+                   :updated-date "2021-07-04T21:55:31Z",
                    :url "https://github.com/AlexFolland/BattlegroundSpiritReleaser"}
                   {:description "AltReps is an addon that allows you to track reputations across your characters",
                    :download-count 0,
@@ -617,7 +617,7 @@
                    :source "github",
                    :source-id "Alastair-Scott/AltReps",
                    :tag-list [],
-                   :updated-date "2021-12-03T00:26:27+00:00",
+                   :updated-date "2021-12-03T00:26:27Z",
                    :url "https://github.com/Alastair-Scott/AltReps"}
                   {:description "Makes system chat messages prettier and tidier, and reduces the need for multiple chat windows."
                    :download-count 0,
@@ -627,7 +627,7 @@
                    :source "github",
                    :source-id "GoldpawsStuff/ChatCleaner",
                    :tag-list [],
-                   :updated-date "2021-12-15T21:17:51+00:00"
+                   :updated-date "2021-12-15T21:17:51Z"
                    :url "https://github.com/GoldpawsStuff/ChatCleaner"}]
         fixture (slurp (fixture-path "github-catalogue--dummy.csv"))
         fake-routes {"https://raw.githubusercontent.com/ogri-la/github-wow-addon-catalogue/main/addons.csv"