+++ test/strongbox/toc_test.clj
@@ -77,6 +77,10 @@ SomeAddon.lua")
                     :#interface "11302"}]
       (is (= expected (toc/parse-toc-file toc-file-contents))))))
 
+(deftest parse-toc-file--empty
+  (testing "empty toc data returns `nil`"
+    (is (nil? (toc/parse-toc-file "")))))
+
 (deftest parse-addon-toc-guard
   (testing "parsing of scraped toc-file key-vals"
     (let [addon-path (join fs/*cwd* "SomeAddon")
@@ -223,3 +227,9 @@ SomeAddon.lua")
         addon-dir (join (helper/install-dir) "EveryAddon")]
     (zip/unzip-file fixture (helper/install-dir))
     (is (= expected (toc/find-toc-files addon-dir)))))
+
+(deftest parse-addon-toc--invalid-toc-questie
+  (testing "invalid toc data is discarded"
+    (let [fixture (helper/fixture-path "questie--invalid.toc")
+          raw-data (toc/read-toc-file fixture)]
+      (is (nil? (toc/parse-addon-toc raw-data))))))